Activity Log: Use server-provided `isDiscarded` value when possible (#21334)
The activity log entries change appearance if they are discarded from
a given perspective. On initial render the perspective is "now" and
the server has precomputed the values. When opening rewind dialogs
though the perspective changes and the values need to be dynamically
created.

In this patch we're defaulting to use the server-provided values to skip
processing time and only calculating the value dynamically when a
different perspective has been chosen.